[cpp05] ex02 트리 색깔 및 srand 수정

+++ cpp/cpp05/ex02/main.cpp
@@ -1,4 +1,4 @@
-#include <time.h>
+#include <ctime>
 
 #include "AForm.hpp"
 #include "Bureaucrat.hpp"
@@ -7,7 +7,7 @@
 #include "ShrubberyCreationForm.hpp"
 
 int main() {
-  std::srand(std::time(NULL));
+  std::srand(static_cast<unsigned int>(std::time(NULL)));
   std::cout << "======문서 목록======" << std::endl;
 
   AForm* shrubbery = new ShrubberyCreationForm("house");
